Update README.md and add prefeitoslaranjalpaulista.md and todosprefeitolaranjaljson
This commit updates the README.md file to include information about the two groups of Prefeitos (Prefeitos Nomeados and Prefeitos Eleitos) that governed Laranjal Paulista, SP. It also adds the prefeitoslaranjalpaulista.md file, which contains a list of all the Prefeitos and their respective periods in office. Additionally, the commit includes the todosprefeitolaranjaljson file, which is a JSON file containing the same information as the prefeitoslaranjalpaulista.md file.